Reorganise packages
- use new core dependencies
- rename packages
